+++ frontend/src/components/Navigation.tsx
@@ -1,14 +1,17 @@
-import { useContext } from 'react';
+import { useContext, useEffect, useRef, useState, MouseEvent } from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { ActionType } from '../actions/Actions';
 import { RequestHelperContext } from '../context/RequestHelperContextProvider';
-import { selectCurrency, selectName, store } from '../store/Store';
+import { selectCurrency, selectUserId, store } from '../store/Store';
+import { Avatar } from './Avatar';
 
 export const Navigation = () => {
   const { setClient } = useContext(RequestHelperContext);
-
+  const userId = useSelector(selectUserId);
   const currency = useSelector(selectCurrency);
+  const [userMenue, setUserMenu] = useState(false);
+  const layerRef = useRef<HTMLDivElement>(null);
 
   const logout = () => {
     setClient!(undefined);
@@ -18,6 +21,32 @@ export const Navigation = () => {
     });
   };
 
+  function handleLinkClick(event: React.MouseEvent<HTMLAnchorElement>) {
+    setUserMenu(false);
+
+    return true;
+  }
+
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent<Document>) {
+      if (
+        layerRef.current &&
+        !layerRef.current.contains(event.target as Node)
+      ) {
+        setUserMenu(false);
+      }
+    }
+
+    const clickHandler: EventListenerOrEventListenerObject =
+      handleClickOutside as unknown as EventListenerOrEventListenerObject;
+
+    document.addEventListener('mousedown', clickHandler);
+
+    return () => {
+      document.removeEventListener('mousedown', clickHandler);
+    };
+  }, [layerRef]);
+
   return (
     <>
       <div className="item">
@@ -25,13 +54,11 @@ export const Navigation = () => {
           <img alt="Deals" src={`${currency?.toLocaleLowerCase()}-icon.svg`} />
         </Link>
       </div>
-
       <div className="item">
         <Link to="/forecast">
           <img alt="Forecast" src="forecast-icon.svg" />
         </Link>
       </div>
-
       <div className="item">
         <Link to="/accounts">
           <img alt="Accounts" src="accounts-icon.svg" />
@@ -46,12 +73,28 @@ export const Navigation = () => {
         <Link to="/setup">
           <img alt="Setup" src="setup-icon.svg" />
         </Link>
-      </div>
-
+      </div>{' '}
       <div className="item" style={{ flexGrow: 1 }}></div>
-
-      <div className="item" onClick={logout}>
-        <img alt="Logout" src="exit-icon.svg" />
+      <div className="user-menu">
+        {userMenue && (
+          <div className={userMenue ? 'wrapper' : ''} ref={layerRef}>
+            <Link
+              onClick={handleLinkClick}
+              to="/user-setup"
+              className="link user-settings"
+            >
+              Settings
+            </Link>
+            <div onClick={logout} className="link logout">
+              Logout
+            </div>
+          </div>
+        )}
+        <Avatar
+          onClick={() => setUserMenu(!userMenue)}
+          width={36}
+          id={userId}
+        />
       </div>
     </>
   );