+++ frontend/src/pages/SetupPage.tsx
@@ -1,12 +1,10 @@
-import { useContext, useEffect, useState } from 'react';
+import { useContext, useEffect } from 'react';
 import { store } from '../store/Store';
 import { ActionType } from '../actions/Actions';
 import { RequestHelperContext } from '../context/RequestHelperContextProvider';
-import { AnimalCanvas } from '../components/setup/AnimalCanvas';
 import { CardCanvas } from '../components/setup/card/CardCanvas';
 import { CurrencyCanvas } from '../components/setup/currency/CurrencyCanvas';
 import { LanesCanvas } from '../components/setup/lane/LaneCanvas';
-import { PasswordCanvas } from '../components/setup/PasswordCanvas';
 
 export const SetupPage = () => {
   const { client } = useContext(RequestHelperContext);
@@ -26,13 +24,6 @@ export const SetupPage = () => {
         type: ActionType.SCHEMAS,
         payload: [...schemas],
       });
-
-      let users = await client!.getUsers();
-
-      store.dispatch({
-        type: ActionType.USERS,
-        payload: [...users],
-      });
     };
 
     if (client) {
@@ -43,14 +34,8 @@ export const SetupPage = () => {
   return (
     <div className="canvas">
       <CurrencyCanvas />
-
       <LanesCanvas />
-
       <CardCanvas />
-
-      <AnimalCanvas />
-
-      <PasswordCanvas />
     </div>
   );
 };